Create Initial Java GUI
threw this together in about 2 hours. VERY simple, but it didn't take much code/time and I think this is how we should start for most of the assignment and build upon it for fancier/nicer looking UI. Took me a while because this is my first experience with Java in a about a year. This code can be cleaned up easily and split into separate classes for easier reading (which I'll try to do tomorrow before class) but this is a start at least of basic functionality.

You all can add your names into the 'About' / 'Who Dun It?' section of the code (around line 95ish I believe)

I just added this file in here manually because I haven't setup my github desktop app to connect my work space to the repository. I will also work on this as well so it can be official? (is there any benefit to not doing it manually? I'll talk with you all about it Thursday if i haven't figured it out by then).