Big Updates!
- Animations are no longer tied to when components are repainted.

This caused a weird bug on pages with GWrappers. Animations would play faster
since JComponents like to repaint themselves a lot.

- GTextBox now uses JTextFields instead of my custom code.

I was rather disappointed with my code since it was missing a lot of basic features
and had some bugs with Windows. Now GTextBox uses a JTextField as a base
and adds on features like the animated error messages. I still have the old boxes
as GTextBoxOLD.

- Buttons now have rounded corners.